Start porting SDL to MyOS.
Create SDL_TestApp1 program, which is meant to be built for both Windows and MyOS with the same source file. Create Two projects for the different build targets. (It may be possible to configure a target platform for MyOS and Windows within the same project but I didn't spend much time attempting this.)

Use a dynamically allocated buffer for executables.

Note: there are some instances where using relative directories was not practical for the test app projects. The test app projects will need to be updated with the include/lib/post-build directories of the target dev machine.