Implement free() to actually free allocated memory. Freed memory is added to an ALLOCATION_ARRAY. Malloc() will check this array for freed memory and reassign freed memory before allocating new memory. This is just a very crude, "first-fit" algorithm that isn't efficient and will definitely lead to wasted memory (but it's better than nothing).
Implement malloc() / free() debugging, which is activated by defining DEBUG_MEM. If this macro is present, the macros dbg_alloc() and dbg_release() will give file name and line number information to malloc() and free(). If DEBUG_MEM is not present, these macros will just call malloc() and free(). The extra debugging info can be viewed for allocations with the "allocations" shell command*. Freeing an invalid pointer with dbg_release() will show the file name and line number of the invalid free(). This can be used to track memory leaks and debug invalid frees().

The new free() implementation and memory debugging has been used to greatly improve the virtio-net driver. The driver now frees allocated memory and several bugs have been fixed. The driver is still not completely bug-free; tftp transmission errors will occur with large transfers. Because of this, sdl_testapp1.exe is not working with the virtio-net driver (and by extension, it's not working with VirtualBox).

*Note: each application keeps track of its own memory allocations with its own ALLOCATION_ARRAY. Only entire pages of memory are allocated by the kernel, with the expectation that (in the future) any pages reserved by an application will be freed by the system when that application exits. So debugging memory leaks with the "allocations" command only makes sense for the kernel and drivers.