Add ExitApp() function for exiting multitasking apps with the exit() system call. Soon ExitApp() will also be the way that apps returning from main() will exit.
Add some code to timer_interrupt_handler() to check if a task has quit, and delete a ready queue entry if so.

Add stack trace to invalid opcode handler.

Expand simple debugger to check the module name to see which map file to load (only the kernel map file is implementing at the moment) and not lookup mappings in unknown modules. However, check the address and if the mapping is in kernel space, then look up the address from the kernel map. Also, reload map files when the kernel has been relaunched.

Add a temporary fix to TFTP to keep track of which task requested the file. If the TFTP receive packet handler is called when another task is running, switch to the page table of the task that requested the file, then switch back when we're done. This is only a hack and doesn't work in all cases. The next step is probably to implement kmalloc() so networking functions and drivers can allocate memory in kernel space without fear of causing a page fault later - this should be a permanent fix when it's implemented.

Remove GUI.map from repo (.map files change too much with each commit and it throws off my line count).